Test day changes

+++ Amiezone/Register.cs
@@ -22,37 +22,51 @@ namespace Amiezone
 
         private void button1_Click(object sender, EventArgs e)
         {
-            //For id just use date and convert to int or long
+            //Need to figure out how to get ticks in right
+            Boolean exists = false;
             DateTime date = new DateTime();
             long n = date.Ticks;
-            string user = UsernameBox.Text + ".txt";
+
+            // Check that everything has been filled in
+            if(UsernameBox.Text == "" || PasswordBox.Text == "" || FundsBox.Text == "" || AddressBox.Text == "")
+            {
+                MessageBox.Show("Please enter your information in every field");
+                return;
+            }
+            // Turns user textbox into a filepath
+            string user = Path.Combine(storeClasses.generalFilePath, "Users", UsernameBox.Text + ".txt");
             string pass = PasswordBox.Text;
             string projectPath = storeClasses.generalFilePath;
             projectPath = Path.Combine(projectPath, "Users");
+
+            // Checks for file existance
             foreach (string itemFile in Directory.EnumerateFiles(projectPath))
             {
                 if (itemFile == user)
                 {
                     MessageBox.Show("User already exists");
+                    exists = true;
                     return;
                 }
             }
-            // Create a file to write to.
 
-            string path = Path.Combine(projectPath, user);
-            string funds = FundsBox.Text;
-            string address = AddressBox.Text;
-            string[] data = { pass, n.ToString(), funds, address };
-            string x = pass + n.ToString() + funds + address;
-            /*using (StreamWriter sw = File.CreateText(path))
+            // Creates a file to write to if file doesn't already exist.
+            if (exists == false)
             {
-                    sw.WriteLine(pass);
-                    sw.WriteLine(n);
-                    sw.WriteLine(funds);
-                    sw.WriteLine(address);
+                // Creates a file to write to.
+                string path = Path.Combine(projectPath, user);
+                string funds = FundsBox.Text;
+                string address = AddressBox.Text;
+                string[] data = { pass, n.ToString(), funds, address };
+
+                System.IO.File.WriteAllLines(path, data);
+                loginPage.Enabled = true;
+                this.Close();
             }
-            */
-            System.IO.File.WriteAllLines(path, data);
+
+        }
+        private void exitRegister(object sender, EventArgs e)
+        {
             loginPage.Enabled = true;
             this.Close();
         }
